requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-ace==1.31.1
django-allauth==0.58.2
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-codemirror==1.0.1
django-codemirror-widget==0.5.0
django-humanize==0.1.2
django-jet-calm==5.1.5
django-js-asset==2.1.0
django-tinymce==3.6.1
djangocodemirror==2.1.0
google-api-python-client==1.4.1
httplib2==0.22.0
humanize==4.9.0
idna==3.4
mysqlclient==2.2.0
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
